backend/note-taking/requirements.txt
flask>=2.3.0
assemblyai>=0.26.0
python-dotenv>=1.0.0
transformers>=4.40.0
huggingface-hub>=0.23.0
torch>=2.0.0
sentencepiece>=0.1.99
gunicorn>=21.2.0
